+++ lib/liboswkeys/Makefile
@@ -25,7 +25,7 @@ LIB=lib${LIBRARY}.a
 MANDIR=$(MANTREE)/man3
 
 ONEFILE=x509dn.c
-SRCS=asn1.c certload.c pem.c pgp.c pkcs.c x509dn.c
+SRCS=asn1.c certload.c pem.c pgp.c pkcs.c x509dn.c ocsp.c
 SRCS+=x509cert.c x509lists.c oid.c x509chain.c
 
 OBJS=${SRCS:.c=.o}
